requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-core>=0.2.0
langgraph>=0.1.0
langsmith>=0.1.0

# LLM Providers (OpenRouter support)
openai>=1.3.0
langchain-openai>=0.1.0

# Vector Database
chromadb>=0.4.22
sentence-transformers>=2.2.2

# API Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# UI
streamlit>=1.28.0

# Evaluation
ragas>=0.1.0
datasets>=2.16.0

# Document Processing
pypdf>=4.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Code Quality
black>=24.1.0
pylint>=3.0.0